requirements.txt
flask-restplus==0.9.2
Flask-SQLAlchemy==2.1
Werkzeug==0.16.1
numpy==1.19.5
pandas==1.1.5
sklearn==0.22.2.post1
OpenCV==4.1.2
tensorflow==2.4.1
keras==2.4.3

# numpy==1.19.5, pandas==1.1.5, sklearn==0.22.2.post1, OpenCV==4.1.2, tensorflow==2.4.1, keras==2.4.3

